Refactor MCP path expression helpers
